+++ packages/calypso-build/webpack/file-loader.js
@@ -1,19 +1,10 @@
 /**
- * Return a Webpack loader configuration object for files / images.
- *
- * @param  {object} options File loader options
- * @returns {object}         Webpack loader object
+ * Return a Webpack loader configuration object for images.
  */
-module.exports.loader = ( options ) => ( {
+module.exports.loader = () => ( {
 	test: /\.(?:gif|jpg|jpeg|png|svg)$/i,
-	use: [
-		{
-			loader: require.resolve( 'file-loader' ),
-			options: {
-				name: '[name]-[hash].[ext]',
-				outputPath: 'images/',
-				...options,
-			},
-		},
-	],
+	type: 'asset/resource',
+	generator: {
+		filename: 'images/[name]-[contenthash][ext]',
+	},
 } );